+++ src/gamspy/_backend/local.py
@@ -32,15 +32,13 @@ from gams import GamsJob, GamsOptions
 from gams.control.workspace import GamsExceptionExecution
 
 import gamspy._backend.backend as backend
+import gamspy._miro as miro
 from gamspy.exceptions import GamspyException, customize_exception
 
 if TYPE_CHECKING:
     import io
 
     from gamspy import Container, Model
-
-MIRO_GDX_IN = os.getenv("GAMS_IDC_GDX_INPUT", None)
-MIRO_GDX_OUT = os.getenv("GAMS_IDC_GDX_OUTPUT", None)
 
 
 class Local(backend.Backend):
@@ -110,14 +108,7 @@ class Local(backend.Backend):
                 self.container._gdx_out, symbols
             )
 
-        if MIRO_GDX_IN and self.container._miro_input_symbols:
-            self.container._load_records_from_gdx(
-                MIRO_GDX_IN, self.container._miro_input_symbols
-            )
-        if MIRO_GDX_OUT and self.container._miro_output_symbols:
-            self.container._load_records_from_gdx(
-                MIRO_GDX_OUT, self.container._miro_output_symbols
-            )
+        miro.load_miro_symbol_records(self.container)
 
         self.container._swap_checkpoints()