+++ src/gamspy/_miro.py
@@ -11,6 +11,44 @@ from gamspy.exceptions import ValidationError
 if TYPE_CHECKING:
     from gamspy import Container, Equation, Parameter, Set, Variable
 
+MIRO_GDX_IN = os.getenv("GAMS_IDC_GDX_INPUT", None)
+MIRO_GDX_OUT = os.getenv("GAMS_IDC_GDX_OUTPUT", None)
+
+
+def get_load_input_str(statement: str, gdx_in: str) -> str:  # pragma: no cover
+    string = "$gdxIn\n"  # close the old one
+    string += f"$gdxIn {MIRO_GDX_IN}\n"  # open the new one
+    string += f"$loadDC {statement}\n"
+    string += "$gdxIn\n"  # close the new one
+    string += f"$gdxIn {gdx_in}\n"
+
+    return string
+
+
+def get_unload_output_str(container: Container) -> str:  # pragma: no cover
+    unload_str = ",".join(container._miro_output_symbols)
+    return f"execute_unload '{MIRO_GDX_OUT}' {unload_str}\n"
+
+
+def load_miro_symbol_records(
+    container: Container, is_implicit: bool = False
+):  # pragma: no cover
+    # Load records of miro input symbols
+    if MIRO_GDX_IN and container._miro_input_symbols:
+        container._load_records_from_gdx(
+            MIRO_GDX_IN, container._miro_input_symbols
+        )
+        for name in container._miro_input_symbols:
+            symbol = container[name]
+            if isinstance(symbol, gp.Parameter) and symbol._is_miro_table:
+                symbol._records.columns = symbol.domain_names + ["value"]
+
+    # Load records of miro output symbols
+    if MIRO_GDX_OUT and container._miro_output_symbols and not is_implicit:
+        container._load_records_from_gdx(
+            MIRO_GDX_OUT, container._miro_output_symbols
+        )
+
 
 class MiroJSONEncoder:
     def __init__(
@@ -23,7 +61,6 @@ class MiroJSONEncoder:
         self.output_symbols = container._miro_output_symbols
         self.input_scalars = self._find_scalars(self.input_symbols)
         self.output_scalars = self._find_scalars(self.output_symbols)
-        self.miro_json = self.prepare_json()
 
     def _find_scalars(self, symbols: list[str]) -> list[str]:
         scalars = []
@@ -141,48 +178,55 @@ class MiroJSONEncoder:
     def prepare_headers_dict(
         self, symbol: Set | Parameter | Variable | Equation
     ):
-        type_map = {
-            str: "string",
-            "float64": "numeric",
-            "category": "string",
-            "object": "string",
-        }
-
-        if isinstance(symbol, gp.Set) and symbol.records is None:
+        if isinstance(symbol, gp.Set):
             domain_keys = ["uni", "element_text"]
             domain_values = [
                 {"type": "string", "alias": "uni"},
                 {"type": "string", "alias": "element_text"},
             ]
             return dict(zip(domain_keys, domain_values))
+        elif isinstance(symbol, gp.Parameter):
+            domain_keys = symbol.domain_names + ["value"]
+            types = ["string"] * len(symbol.domain_names) + ["numeric"]
 
-        domain_keys = symbol.records.columns.to_list()
-        domain_values = []
+            if symbol._is_miro_table:
+                last_item = symbol.domain[-1]
+                self.validate_table(symbol, last_item)
 
-        for dtype, column in zip(symbol.records.dtypes, domain_keys):
-            try:
-                elem = self.container[column]
-                alias = elem.description if elem.description else elem.name
-            except KeyError:
-                alias = column
-
-            domain_values.append(
-                {"type": type_map[dtype.name], "alias": alias}
-            )
-
-        if isinstance(symbol, gp.Parameter) and symbol._is_miro_table:
-            last_item = symbol.domain[-1]
-            self.validate_table(symbol, last_item)
-
-            if isinstance(last_item, (gp.Set, gp.Alias)):
                 set_values = last_item.records["uni"].values.tolist()
 
                 domain_keys = domain_keys[:-2]
+                types = ["string"] * len(domain_keys) + ["numeric"] * len(
+                    set_values
+                )
                 domain_keys += set_values
+        elif isinstance(symbol, (gp.Variable, gp.Equation)):
+            domain_keys = symbol.domain_names + [
+                "level",
+                "marginal",
+                "lower",
+                "upper",
+                "scale",
+            ]
+            types = ["string"] * len(symbol.domain_names) + ["numeric"] * 5
 
-                domain_values = domain_values[:-2]
-                for elem in last_item.records["uni"].values.tolist():
-                    domain_values.append({"type": "numeric", "alias": elem})
+        domain_values = []
+        uni_counter = 0
+        for idx, key in enumerate(domain_keys):
+            if key == "*":
+                domain_keys[idx] = (
+                    "uni" if uni_counter == 0 else f"uni{uni_counter}"
+                )
+                uni_counter += 1
+
+        for column, column_type in zip(domain_keys, types):
+            try:
+                elem = self.container[column]
+                alias = elem.description if elem.description else column
+            except KeyError:
+                alias = column
+
+            domain_values.append({"type": column_type, "alias": alias})
 
         assert len(domain_keys) == len(domain_values)
         return dict(zip(domain_keys, domain_values))
@@ -250,7 +294,7 @@ class MiroJSONEncoder:
 
         return symbols_dict
 
-    def prepare_json(self) -> str:
+    def prepare_dict(self) -> dict:
         input_symbols_dict = self.prepare_symbols_dict(
             True, self.input_symbols
         )
@@ -264,10 +308,10 @@ class MiroJSONEncoder:
             "outputSymbols": output_symbols_dict,
         }
 
-        return json.dumps(miro_dict, indent=4)
+        return miro_dict
 
-    def writeJson(self):
-        content = self.prepare_json()
+    def write_json(self):
+        miro_dict = self.prepare_dict()
 
         filename = os.path.splitext(os.path.basename(sys.argv[0]))[0]
         directory = os.path.dirname(sys.argv[0])
@@ -278,4 +322,6 @@ class MiroJSONEncoder:
             pass
 
         with open(os.path.join(conf_path, f"{filename}_io.json"), "w") as conf:
-            conf.write(content)
+            conf.write(json.dumps(miro_dict, indent=4))
+
+        return miro_dict