+++ src/gamspy/_symbols/alias.py
@@ -34,13 +34,12 @@ import gamspy._algebra.condition as condition
 import gamspy._algebra.operable as operable
 import gamspy._validation as validation
 from gamspy._symbols.set import SetMixin
-from gamspy._symbols.symbol import Symbol
 
 if TYPE_CHECKING:
     from gamspy import Container, Set
 
 
-class Alias(gt.Alias, operable.Operable, Symbol, SetMixin):
+class Alias(gt.Alias, operable.Operable, SetMixin):
     """
     Represents an Alias symbol in GAMS.
     https://www.gams.com/latest/docs/UG_SetDefinition.html#UG_SetDefinition_TheAliasStatementMultipleNamesForASet