+++ src/gamspy/_symbols/equation.py
@@ -573,6 +573,7 @@ class Equation(gt.Equation, operable.Operable, Symbol):
 
         if self._records is not None and self.domain_forwarding:
             self._domainForwarding()
+            self._mark_forwarded_domain_sets()
 
             # reset state check flags for all symbols in the container
             for symbol in self.container.data.values():