+++ src/gamspy/_symbols/parameter.py
@@ -333,6 +333,7 @@ class Parameter(gt.Parameter, operable.Operable, Symbol):
 
         if self._records is not None and self.domain_forwarding:
             self._domainForwarding()
+            self._mark_forwarded_domain_sets()
 
             # reset state check flags for all symbols in the container
             for symbol in self.container.data.values():